requirements.txt
requests
pandas
python-dotenv
numpy==1.23.5
pandas_ta==0.3.14b0